requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
revChatGPT==3.0.9
PyNaCl==1.5.0
